add-email functionality.
